requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.18
pypdf>=3.17.1
python-docx>=1.1.0
docx2txt>=0.8
unstructured>=0.11.6
markdown>=3.5.1
python-dotenv>=1.0.0
pydantic>=2.5.0